+++ hubdle-gradle-plugin/testFunctional/resources/config/coverage/1/library/main/kotlin/b.kt
@@ -1,0 +1,5 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.kotlin.jvm.sandbox.project.library
+
+fun b() = "B"