+++ hubdle-gradle-plugin/testFunctional/resources/config/coverage/1/library/test/kotlin/aTest.kt
@@ -1,0 +1,13 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.kotlin.jvm.sandbox.project.library
+
+import kotlin.test.Test
+
+class BTest {
+
+    @Test
+    fun bTest() {
+        require(b() == "B")
+    }
+}