+++ hubdle-gradle-plugin/testFunctional/resources/config/coverage/1/main/kotlin/a.kt
@@ -1,0 +1,5 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.kotlin.jvm.sandbox.project
+
+fun a() = "A"