+++ hubdle-gradle-plugin/testFunctional/resources/config/coverage/1/test/kotlin/aTest.kt
@@ -1,0 +1,13 @@
+@file:Suppress("PackageDirectoryMismatch")
+
+package com.kotlin.jvm.sandbox.project
+
+import kotlin.test.Test
+
+class ATest {
+
+    @Test
+    fun aTest() {
+        require(a() == "A")
+    }
+}